Move method to rest api controller.
